fix: pin black version and fix formatting
A new version of black was recently pulled into Travis builds
and caused some lint-related errors because of some slight
differences in the code formatting output.
This commit pins the version of black used in our builds
to v24.x.x and fixes the handful of existing formatting-related
problems identified by this new version of black.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>